ice_dyn_shared: add optional 'grid_location' argument to seabed_stress_factor_LKD
In subsequent commits we want to compute the seabed stress factor
'Tb' not only at the U point (NE corner) but also at the E (center of
E face) and N (center of N face) points.

In order to dispatch the computation in this subroutine to different
code paths depending on the grid location (U, N or E), add an optional
argument 'grid_location' that can be used to indicate at which point we
want the factor to be computed. Default it to the 'U' point for
backwards compatibility, such that the existing calls do not have to be
changed.

Note that we need an additional 'l_grid_location' variable to set the
default value, as the Fortran standard does not allow setting the value
of an optional argument if it is not present.

Note also that 'icellu' was incorrectly referencing 'icetmask'. Fix that
by generalizing the description.