ice_dyn_shared: use 'grid_neighbor_{min,max}' in seabed_stress_factor_LKD
Use the 'grid_neighbor_{min,max}' subroutines introduced in the
preceding commit to compute the min and max of neighbor values of
hwater, aice and vice at 'grid_location'. This generalizes the
subroutine for the CD-grid. A subsequent commit will call it twice to
compute the seabed stress factor at the E and N locations.