ice_dyn_shared: add optional CD-grid arguments to 'seabed_stress_factor_prob'
In contrast to 'seabed_stress_factor_LKD', in
'seabed_stress_factor_prob' the seabed stress factor is computed at the
T location and only transfered to the U location at the end of the
subroutine. So for efficiency it does not make sense to call that
subroutine twice for computing the factor at the E and N locations.

Instead, add optional arguments for 'TbE' and 'TbN', as well as the
corresponding indices paraphernalia, and compute the factor at the E and
N location at the end of the subroutine depending on 'grid_system' and
the presence of the optional arguments.

Use 'grid_neighbor_max' to abstract away the operation of finding the
maximum value of neighboring cells.