ice_dyn_shared: add 'step_vel' to step velocities at CD-grid locations
Add a new subroutine 'step_vel', similar to the existing 'stepu', to
step the CD-grid velocities forward by solving the momentum equation.

In contrast to 'stepu', the components of the divergence of the internal
stress tensor are computed in a separate subroutine, 'div_stress', and
so are simply passed as arguments in 'step_vel'.

We mostly follow 'stepu' for the rest of the computation, apart from two
micro-optimizations:

- Introduce a new intermediate variable, 'ccc', to compute the
denominator of the seabed stress coefficient, and reuse it when
computing the seabed stress on the last iteration.

- Fold the two 'if's at the end of the subroutine into a single one.