ice_dyn_shared: generalize 'principal_stress' arguments names
In a subsequent commit we will call 'principal_stress' with the CD-grid
stress arrays 'stress{p,m,12}T' to compute the principal stresses at the
tracer point when using the CD grid.

In that light, remove the '_1' suffix from the stress arguments since
they won't always be located at the NE corner anymore.