ice_history: compute 'sig[12P]' at T-point for CD-grid
For the CD-grid, compute the principal stresses and the ice pressure at
the T point by passing the appropriate arrays to 'principal_stress'.

These three history variables are computed using the NE-corner values on
the B-grid, but this is not reflected in the description of the fields
in the history output.

Add the location at which the stresses are computed to the comment
argument in the call to 'define_hist_field', for both the B and CD grid.