ice_dyn_evp: pass 'stress{p,m,12}[TU]' to dyn_prep2
For the CD-grid, the 'stress{p,m,12}_[1-4]' arrays are not used. Pass
the CD-grid location stress arrays 'stress{p,m,12}[TU]' to 'dyn_prep2',
which zero-initializes them anywhere icetmask is zero.

Also, add a TODO in the `if (grid_type) == 'tripole'` block, since it is
not yet clear how the halo updates should be done for the new stress
arrays on the tripole grid.